NFC FUNCIONA i llegeix Long id perfectament!
Falta:
  -Que es vegi be quan un producte te més de 2 xifres al preu o cantitat
  -Implementar updateAllProducts realment

+++ app/src/main/java/xyz/carlesllobet/pushmarket/DB/DatabaseHandler.java
@@ -61,7 +61,7 @@ public class DatabaseHandler extends SQLiteOpenHelper {
     /**
      * Storing user details in database
      * */
-    public boolean addProduct(Integer id, String nombre, String description, Integer sector, Integer preu, Uri pic) {
+    public boolean addProduct(Long id, String nombre, String description, Integer sector, Integer preu, Uri pic) {
     	SQLiteDatabase db = this.getWritableDatabase();
 
         //Si existeix, retorna fals, i no es pot afegir
@@ -99,7 +99,7 @@ public class DatabaseHandler extends SQLiteOpenHelper {
         // Move to first row
         if (cursor.moveToFirst()) {
             do {
-                Product aux = new Product(cursor.getInt(0),cursor.getString(1),cursor.getString(2),Uri.parse(cursor.getString(3)),
+                Product aux = new Product(cursor.getLong(0),cursor.getString(1),cursor.getString(2),Uri.parse(cursor.getString(3)),
                         cursor.getInt(4),cursor.getInt(5));
                 products.add(aux);
             } while (cursor.moveToNext());
@@ -109,7 +109,7 @@ public class DatabaseHandler extends SQLiteOpenHelper {
         return products;
     }
 
-    public Boolean CheckProductExist(Integer id) {
+    public Boolean CheckProductExist(Long id) {
         SQLiteDatabase db = this.getReadableDatabase();
         String selectQuery="SELECT * FROM " + TABLE_PRODUCTS + " WHERE " + KEY_ID + " = " + id;
         Cursor c = db.rawQuery(selectQuery, null);
@@ -131,14 +131,14 @@ public class DatabaseHandler extends SQLiteOpenHelper {
         return "";
     }
 
-    public Product getProduct (Integer id) {
+    public Product getProduct (Long id) {
         SQLiteDatabase db = this.getReadableDatabase();
         String selectQuery="SELECT * FROM " + TABLE_PRODUCTS + " WHERE " + KEY_ID + " = " + id;
         Cursor c = db.rawQuery(selectQuery, null);
         Product p = null;
         if (c.moveToFirst()) {
             db.close();
-            p = new Product(c.getInt(0),c.getString(1),c.getString(2),Uri.parse(c.getString(3)),c.getInt(4),c.getInt(5));
+            p = new Product(c.getLong(0),c.getString(1),c.getString(2),Uri.parse(c.getString(3)),c.getInt(4),c.getInt(5));
         }
         db.close();
         return p;
