+++ app/src/main/java/xyz/carlesllobet/pushmarket/Domain/Product.java
@@ -8,7 +8,7 @@ import android.net.Uri;
 public class Product {
 
     //private variables
-    Integer id;
+    Long id;
     String name;
     String description;
     Integer sector;
@@ -21,7 +21,7 @@ public class Product {
     }
 
     // constructor
-    public Product(Integer id, String name, String description, Uri pic, Integer sector, Integer preu) {
+    public Product(Long id, String name, String description, Uri pic, Integer sector, Integer preu) {
         this.id = id;
         this.name = name;
         this.description = description;
@@ -30,9 +30,9 @@ public class Product {
         this.pic = pic;
     }
 
-    public Integer getId() {return this.id;}
+    public Long getId() {return this.id;}
 
-    public void setId(Integer id) {this.id = id;}
+    public void setId(Long id) {this.id = id;}
 
     // getting name
     public String getName() {
